Add exceptions StageController

+++ s15-10/demo/src/main/java/s1510/demo/controller/StageController.java
@@ -2,12 +2,15 @@ package s1510.demo.controller;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.springframework.dao.DataAccessException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.util.UriComponentsBuilder;
 import s1510.demo.dtos.request.StageRequest;
 import s1510.demo.dtos.response.StageResponse;
+import s1510.demo.exception.BadRequestException;
+import s1510.demo.exception.ResourceNotFoundException;
 import s1510.demo.service.StageService;
 
 import java.net.URI;
@@ -22,22 +25,31 @@ public class StageController {
 
     @GetMapping("/stages")
     public ResponseEntity<List<StageResponse>> listStages() {
+        if (stageService.listAll().isEmpty()) throw new ResourceNotFoundException("No hay registros de Stages en el sistema");
         return ResponseEntity.ok(stageService.listAll());
     }
 
     @GetMapping("/stage/{id}")
     public ResponseEntity<Optional<StageResponse>> getStage(@PathVariable Long id) {
+        if (stageService.findById(id) == null) throw new ResourceNotFoundException(String.format("Stage no fue encontrado con: id = '%s'", id));
         return ResponseEntity.ok(stageService.findById(id));
     }
 
     @DeleteMapping("/stage/{id}")
     public ResponseEntity<StageResponse> deleteStage(@PathVariable Long id) {
-        StageResponse stageResponse = stageService.delete(id);
-        return new ResponseEntity<>(stageResponse, HttpStatus.NO_CONTENT);
+        try {
+            if (!stageService.existsById(id)) throw new ResourceNotFoundException(String.format("Stage no fue encontrado con: id = '%s'", id));
+            StageResponse stageResponse = stageService.delete(id);
+            return new ResponseEntity<>(stageResponse, HttpStatus.NO_CONTENT);
+        } catch (DataAccessException e) {
+            throw new BadRequestException(e.getMessage());
+        }
+
     }
 
     @PostMapping("/stage")
     public ResponseEntity<StageResponse> createStage(@RequestBody @Valid StageRequest stageRequest, UriComponentsBuilder uriComponentsBuilder) {
+        if (!stageService.existsById(stageRequest.id())) throw new ResourceNotFoundException(String.format("Stage no fue encontrado con: id = '%s'", stageRequest.id()));
         StageResponse stageResponse = stageService.save(stageRequest);
         URI url = uriComponentsBuilder.path("api/v1/stage/{id}").buildAndExpand(stageResponse.id()).toUri();
         return ResponseEntity.created(url).body(stageResponse);
@@ -49,7 +61,7 @@ public class StageController {
             StageResponse stageResponse = stageService.save(stageRequest);
             return ResponseEntity.ok(stageResponse);
         } else {
-            return ResponseEntity.notFound().build();
+            throw new ResourceNotFoundException(String.format("Stage no fue encontrado con: id = '%s'", stageRequest.id()));
         }
     }
 }
